+++ src/models/index.ts
@@ -8,12 +8,6 @@ import SolidDocument from './SolidDocument';
 import SolidTypeIndex from './SolidTypeIndex';
 import SolidTypeRegistration from './SolidTypeRegistration';
 
-import type DeletesModels from './mixins/DeletesModels';
-import type ManagesPermissions from './mixins/ManagesPermissions';
-import type SerializesToJsonLD from './mixins/SerializesToJsonLD';
-import type { PermissionsTracker } from './mixins/ManagesPermissions';
-import type { SerializeOptions } from './mixins/SerializesToJsonLD';
-
 export * from './helpers';
 export * from './history/index';
 export * from './inference';
@@ -37,10 +31,16 @@ export { default as SolidDocumentSchema } from './SolidDocument.schema';
 export { default as SolidTypeIndexSchema } from './SolidTypeIndex.schema';
 export { default as SolidTypeRegistrationSchema } from './SolidTypeRegistration.schema';
 
-export type { DeletesModels, SerializesToJsonLD, ManagesPermissions, PermissionsTracker, SerializeOptions };
+export type { default as DeletesModels } from './mixins/DeletesModels';
+export type { default as ManagesPermissions } from './mixins/ManagesPermissions';
+export type { default as MigratesSchemas } from './mixins/MigratesSchemas';
+export type { default as SerializesToJsonLD } from './mixins/SerializesToJsonLD';
+export type { PermissionsTracker } from './mixins/ManagesPermissions';
+export type { SerializeOptions } from './mixins/SerializesToJsonLD';
 
 export type { This as DeletesModelsThis } from './mixins/DeletesModels';
 export type { This as ManagesPermissionsThis } from './mixins/ManagesPermissions';
+export type { This as MigratesSchemasThis } from './mixins/MigratesSchemas';
 export type { This as SerializesToJsonLDThis } from './mixins/SerializesToJsonLD';
 
 export type {