+++ src/solid/constants.ts
@@ -1,5 +1,7 @@
 import IRI from '@/solid/utils/IRI';
 
+export const CRDT_PROPERTY = IRI('crdt:property');
+export const CRDT_RESOURCE = IRI('crdt:resource');
 export const LDP_CONTAINER = IRI('ldp:Container');
 export const LDP_CONTAINS = IRI('ldp:contains');
 export const RDF_TYPE = IRI('rdf:type');