+++ src/solid/operations/ChangeUrlOperation.ts
@@ -1,9 +1,9 @@
-import { OperationType } from './Operation';
+import { OperationTypes } from './Operation';
 import type Operation from './Operation';
 
 export default class ChangeUrlOperation implements Operation {
 
-    public type: OperationType.ChangeUrl = OperationType.ChangeUrl;
+    public type: typeof OperationTypes.ChangeUrl = OperationTypes.ChangeUrl;
 
     constructor(public resourceUrl: string, public newResourceUrl: string) {}