+++ src/solid/operations/RemovePropertyOperation.ts
@@ -1,9 +1,9 @@
-import { OperationType } from './Operation';
+import { OperationTypes } from './Operation';
 import type Operation from './Operation';
 
 export default class RemovePropertyOperation implements Operation {
 
-    public type: OperationType.RemoveProperty = OperationType.RemoveProperty;
+    public type: typeof OperationTypes.RemoveProperty = OperationTypes.RemoveProperty;
 
     constructor(public resourceUrl: string, public property?: string, public value?: unknown) {}